Moved the getChatRoomsPosts function to a seperate file (chatRoomsQueries.js). Reference #12.
